requirements.txt
albumentations==1.3.1
numpy==1.24.2
Pillow==9.4.0
Pillow==10.0.0
torch==2.0.1+cu118
torchvision==0.15.2
tqdm==4.65.0
